% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quran.R
\docType{package}
\name{quRan}
\alias{quRan}
\alias{quRan-package}
\alias{quRan-package}
\title{The Qur'an}
\description{
This package contains the complete text of the Qur'an in Arabic (with and
without vowels) and in English (the Yusuf Ali and Saheeh International
translations), formatted to be convenient for text analysis.
}
